+++ changes.md
@@ -106,6 +106,13 @@
 
 
 <details>
+<summary>Removed Resource(s)</summary>
+
+- :warning: removed resource `/{projectKey}/me/carts/key={key}`
+</details>
+
+
+<details>
 <summary>Removed QueryParameter(s)</summary>
 
 - :warning: removed query parameter `localeProjection` from method `get /{projectKey}/products`
@@ -171,6 +178,12 @@
 <summary>Removed Method(s)</summary>
 
 - :warning: removed method `$apiRoot->withProjectKey()->products()->search()->head()`
+- :warning: removed method `$apiRoot->withProjectKey()->me()->carts()->withKey()->get()`
+- :warning: removed method `$apiRoot->withProjectKey()->me()->carts()->withKey()->head()`
+- :warning: removed method `$apiRoot->withProjectKey()->me()->carts()->withKey()->post()`
+- :warning: removed method `$apiRoot->withProjectKey()->me()->carts()->withKey()->delete()`
+- :warning: removed method `$apiRoot->withProjectKey()->me()->quoteRequests()->withId()->delete()`
+- :warning: removed method `$apiRoot->withProjectKey()->me()->quoteRequests()->withKey()->delete()`
 </details>